+++ NEWS.md
@@ -1,5 +1,11 @@
 # shinyTools
 
+## Version 24.03.4
+
+### New Features
+- module for plotExport (#6)
+  - option to set ranges of a plot before export if the plot is a ggplot
+
 ## Version 24.03.3
 
 ### New Features